Annotate percent mitochondrial reads per cell

Add a Percentage of Mitochondrial Read Count Categorical Variable to the Object (based on nCount_RNA)

Usage

add_percent_mito(object, experiment = "gene")

Arguments

object

A object

experiment

gene

Value

a single cell object with cell metadata column containing mitochondrial percentage

Examples

data(small_example_dataset)
add_percent_mito(small_example_dataset)
#> class: SingleCellExperiment 
#> dim: 1000 200 
#> metadata(1): markers
#> assays(2): counts logcounts
#> rownames(1000): Gene_0001 Gene_0002 ... Gene_0999 Gene_1000
#> rowData names(0):
#> colnames(200): Cell_001 Cell_002 ... Cell_199 Cell_200
#> colData names(18): Mutation_Status Cell_Cycle ...
#>   altexps_Spikes_percent total
#> reducedDimNames(3): PCA TSNE UMAP
#> mainExpName: gene
#> altExpNames(1): Spikes