Run SingleCellExperiment Pipeline

This functions allows you to preprocess, cluster and reduce dimensions for one SingleCellExperiment object.

Usage

sce_process(
  object,
  experiment = "gene",
  resolution = 0.6,
  reduction = "PCA",
  organism = "human",
  ...
)

Arguments

object

A SingleCellExperiment object

experiment

Assay of interest in SingleCellExperiment object

resolution

Resolution for clustering cells. Default set to 0.6.

reduction

Dimensional reduction object

organism

Organism

...

extra parameters passed to internal functions

Value

a processed SingleCellExperiment object

Examples

data(small_example_dataset)
sce_process(small_example_dataset)
#> Warning: more singular values/vectors requested than available
#> Warning: You're computing too large a percentage of total singular values, use a standard svd instead.
#> Warning: did not converge--results might be invalid!; try increasing work or maxit
#> [21:51:13] Clustering Cells...
#> clustering at 0.6 resolution
#> stashing markers for gene_snn_res.1
#> class: SingleCellExperiment 
#> dim: 1000 200 
#> metadata(1): markers
#> assays(2): counts logcounts
#> rownames(1000): Gene_0001 Gene_0002 ... Gene_0999 Gene_1000
#> rowData names(1): subsets_Mito
#> colnames(200): Cell_001 Cell_002 ... Cell_199 Cell_200
#> colData names(22): Mutation_Status Cell_Cycle ...
#>   altexps_Spikes_percent total
#> reducedDimNames(3): PCA TSNE UMAP
#> mainExpName: gene
#> altExpNames(1): Spikes